extracting gen-key to a script and removing some useless code
