+++ bin/agent-scripts/generateSshKeyPair
@@ -1,0 +1,10 @@
+#!/bin/sh
+key_name=$1
+ssh-keygen -t rsa -b 1024 -f $key_name -q -N ""
+private_key=`cat $key_name`
+public_key=`cat $key_name.pub`
+
+echo $private_key
+echo "KEY SEPARATOR"
+echo $public_key
+rm $key_name $key_name.pub
